呼び出し側の書き換え(DraftEditPageはServer Component化した)
