+++ src/features/drafts/components/client/DraftEditForm/index.ts
@@ -1,0 +1,1 @@
+export { DraftEditForm } from './DraftEditForm';
